==> AnalysisScript/P1_08_TrainRFModel.R <==
library(raster)
library(sf)
library(dplyr)
library(data.table)
library(reshape2)
library(caret)
library(randomForest)
library(foreach)
library(doParallel)
library(kit)
library(groupdata2)
library(e1071)
library(ranger)
library(matrixStats)
library(TTR)
library(tuneRanger)
library(blockCV)

# Prepare data

setwd("D:/landsat/train")
imgdf <- fread("imgdata_all_t.csv")
trainpt <- read.csv("trainpt.csv")
sb1 <- readRDS("spatialblock.rds")

trainpt_m <- reshape2::melt(trainpt[,c(1,3:14)], id.vars = c("PointID"),
                            variable.name = "yearclass", value.name = "Class")
trainpt_m$Class <- factor(trainpt_m$Class, levels = c(1:6))
levels(trainpt_m$Class) <- c("c1", "c2", "c3", "c4", "c5", "c6")
trainpt_m$fold <- sb1$folds_ids
traindata <- merge(imgdf, trainpt_m, by = c("PointID","yearclass"))
traindata$row_num <- seq.int(nrow(traindata))

predictors <- c("G","R","NIR","ndvi","gndvi","glcm_mean","glcm_con","Slope","Northness")


# Function to build RF model and perform 10-fold cross validation

buildmodel <- function(traindata, predictors, ncore=4, seed=2022){
  set.seed(seed)
  seeds <- sample.int(n=2000, 100)

  # Function to convert raw predictions to probabilities
  rawtoprob <- function(rawpred){
    rawpred <- rawpred$predictions
    df <- data.frame(matrix(ncol = 6, nrow = nrow(rawpred)))
    for (n in 1:6){df[,n] <- apply(rawpred==n, 1, sum)}
    df <- df/sum(df[1,])
    colnames(df) <- c("c1", "c2", "c3", "c4", "c5", "c6")
    return(df)
  }

  # Tune optimal mtry
  td_t <- as.data.table(balance(traindata, "min", cat_col="Class"))
  res <- data.frame()
  for (s in 1:5){
    set.seed(seeds[s])
    res1 <- tuneMtryFast(Class ~., data=td_t[,c(..predictors1,"Class")], stepFactor=1.5, num.threads = ncore)
    res1 <- as.data.frame(res1)
    res <- rbind(res, res1)      
  }
  res_mtry <- res %>% group_by(mtry) %>% summarize(error=mean(OOBError))
  res_mtry <- res_mtry[order(res_mtry$error),]
  bestmtry <- res_mtry$mtry[1]
  print(paste0("bestmtry: ", bestmtry))
    
  # Function to train and predict each fold
  doranger_f <- function(f){
    td <- traindata[traindata$fold != f,]
    pd <- traindata[traindata$fold == f,]
    set.seed(seeds[5]+f)
    td_b <- as.data.table(balance(td, "max", cat_col="Class"))
    set.seed(seeds[6]+f)
    m <- ranger(x = td_b[,..predictors1], y = td_b[["Class"]], mtry = bestmtry,
                seed = seeds[6]+f, num.threads = ncore, verbose = TRUE)
    set.seed(seeds[7]+f)
    m_pred <- predict(m, data = pd, predict.all = TRUE, 
                      seed = seeds[7]+f, num.threads = ncore, verbose = TRUE)
    m_pred <- rawtoprob(m_pred)
    m_pred <- data.table(pd[,c("PointID","yearclass")], repeats=1, m_pred)
    return(m_pred)
  }
    
  # Loop through each fold to obtain cross-validated predictions
  cv_pred <- data.table()
  for (f in 1:length(unique(traindata$fold))){
    print(paste0("Start fold: ", f))
    m_pred <- doranger_f(f)
    cv_pred <- rbind(cv_pred, m_pred)
  }
  
  # Final model using all data
  set.seed(seeds[8])
  td_b <- as.data.table(balance(traindata, "max", cat_col="Class"))
  set.seed(seeds[9]+f)
  finalmodel <- ranger(x = td_b[,..predictors1], y = td_b[["Class"]], 
                       mtry = bestmtry, importance = "impurity", 
                       seed = seeds[9]+f, num.threads = ncore, verbose = TRUE)
  
  return(list(mtry=bestmtry, cv_pred=cv_pred, finalmodel=finalmodel))
}


# Function to combine votes and perform temporal smoothing

voteresult <- function(pred_df, temporalsmooth = TRUE){
  if (inherits(pred_df, "list")){pred_df <- pred_df[[1]]}
  # combine vote
  pred_df_vote <- pred_df %>%
    group_by(PointID, yearclass, repeats) %>%
    summarise(s1=sum(c1),s2=sum(c2),s3=sum(c3),s4=sum(c4),s5=sum(c5),s6=sum(c6))
  pred_df_vote$probsum <- apply(pred_df_vote[4:9], 1, sum)
  for (i in 4:9){pred_df_vote[,i] <- pred_df_vote[,i]/pred_df_vote$probsum}
  pred_df_vote <- as.data.table(pred_df_vote[,1:9])
  # temporal smoothing
  if (temporalsmooth==TRUE){
    tempsmooth <- function(pd_df){
      pd_df_smooth <- data.table()
      for (i in unique(pd_df$PointID)){
        d1 <- pd_df[pd_df$PointID==i,]
        if (nrow(d1) < 10){
          allyear <- paste0("c",seq(1975,2020,5))
          missingyear <- allyear[!allyear %in% d1$yearclass]
          for (m in missingyear){
            d1 <- rbind(d1, data.table(PointID=i, yearclass=m, repeats=r,
                                       s1=0,s2=0,s3=0,s4=0,s5=0,s6=0))
          }
        }
        d1 <- d1[order(d1$yearclass),]
        d1_dec <- d1[order(d1$yearclass, decreasing = TRUE),]
        p1 <- apply(d1_dec[,4:9], 2, WMA, n=2, wts=c(1,2))[10,, drop=FALSE]
        p2 <- apply(d1[,4:9], 2, WMA, n=3, wts=c(1,2,1))[3:10,, drop=FALSE]
        p3 <- apply(d1[,4:9], 2, WMA, n=2, wts=c(1,2))[10,, drop=FALSE]
        d1 <- cbind(d1[,1:3], rbind(p1,p2,p3))
        pd_df_smooth <- rbind(pd_df_smooth, d1)
      }
      return(pd_df_smooth)
    }
    pred_df_vote_smooth <- data.table()
    for (r in unique(pred_df_vote$repeats)){
      pred_df_vote_s1 <- pred_df_vote[pred_df_vote$repeats==r,]
      pred_df_vote_s1 <- tempsmooth(pred_df_vote_s1)
      pred_df_vote_smooth <- rbind(pred_df_vote_smooth, pred_df_vote_s1)
    }
    pred_df_vote <- pred_df_vote_smooth
  }
  # Get class
  pred_df_vote$predclass <- apply(pred_df_vote[,4:9], 1, which.max)
  pred_df_vote$predclass <- factor(pred_df_vote$predclass, levels = c(1:6))
  levels(pred_df_vote$predclass) <- c("c1", "c2", "c3", "c4", "c5", "c6")
  # join ref class
  trainpt_m <- reshape2::melt(trainpt[,c(1,5:14)], id.vars = c("PointID"),
                              variable.name = "yearclass", value.name = "refclass")
  trainpt_m$refclass <- factor(trainpt_m$refclass, levels = c(1:6))
  levels(trainpt_m$refclass) <- c("c1", "c2", "c3", "c4", "c5", "c6")
  pred_df_vote <- merge(pred_df_vote, trainpt_m, by = c("PointID","yearclass"))
  return(pred_df_vote)
}

# Function to combine accuracy and confidence intervals

doaccuracy <- function(pred_df_vote){
  # accuracy assessment
  # https://blogs.fu-berlin.de/reseda/area-adjusted-accuracies/
  # area adjusted accuracy assessment (as in Olofsson et al. 2014)
  confmat <- table(pred_df_vote$predclass, pred_df_vote$refclass)/5
  nclass <- length(unique(pred_df_vote$refclass))
  mappixels <- table(pred_df_vote[pred_df_vote$PointID<=825,]$predclass)
  maparea <- as.numeric(mappixels/sum(mappixels))*1117
  conf <- 1.96
  A <- sum(maparea)
  W_i <- maparea / A
  n_i <- rowSums(confmat) 
  p <- W_i * confmat / n_i
  p[is.na(p)] <- 0
  PA <- diag(p) / colSums(p)
  UA <- diag(p) / rowSums(p)
  OA_CI <- conf * sqrt(sum(W_i ^ 2 * UA * (1 - UA) / (n_i - 1)))
  UA_CI <- conf * sqrt(UA * (1 - UA) / (n_i - 1)) 
  N_j <- sapply(1:nclass, function(x) sum(maparea / n_i * confmat[ , x]) )
  tmp <- sapply(1:nclass, function(x) sum(maparea[-x] ^ 2 * confmat[-x, x] / n_i[-x] * ( 1 - confmat[-x, x] / n_i[-x]) / (n_i[-x] - 1)) )
  PA_CI <- conf * sqrt(1 / N_j ^ 2 * (maparea ^ 2 * ( 1 - PA ) ^ 2 * UA * (1 - UA) / (n_i - 1) + PA ^ 2 * tmp))
  OA <- sum(diag(confmat))/sum(confmat)
  PA <- diag(confmat)/colSums(confmat)
  UA <- diag(confmat)/rowSums(confmat)
  # kappa
  diagonal.counts <- diag(confmat)
  N <- sum(confmat)
  row.marginal.props <- rowSums(confmat)/N
  col.marginal.props <- colSums(confmat)/N
  Po <- sum(diagonal.counts)/N
  Pe <- sum(row.marginal.props*col.marginal.props)
  ka <- (Po - Pe)/(1 - Pe)
  
  pred_df_a <- data.frame(class=c("c1", "c2", "c3", "c4", "c5", "c6"),
                          pa=PA, pa_ci=PA_CI, ua=UA, ua_ci=UA_CI)
  pred_df_a <- rbind(pred_df_a, data.frame(class="oa_ka", pa=OA, pa_ci=OA_CI,
                                           ua=ka, ua_ci=0))
  return(pred_df_a)
}

# Function to combine accuracy and confidence intervals (for individual period)

doaccuracy_sepyear <- function(pred_df_vote){
  yearclass <- unique(pred_df_vote$yearclass)
  pred_df_a <- data.frame()
  for (y in yearclass){
    pred_df_y <- pred_df_vote[pred_df_vote$yearclass==y,]
    pred_df_a1 <- doaccuracy(pred_df_y)
    pred_df_a1$yearclass <- y
    pred_df_a <- rbind(pred_df_a, pred_df_a1)
  }
  return(pred_df_a)
}


# Execute functions

cv_pred <- buildmodel(traindata, predictors)
saveRDS(cv_pred, "D:/landsat/train/cv_pred.rds")
pred_df <- cv_pred[[2]]

pred_df_vote <- voteresult(pred_df)
vresult <- doaccuracy(pred_df_vote)
vresult_sepyear <- doaccuracy_sepyear(pred_df_vote)
write.csv(pred_df_vote,"pred_df_vote.csv")
write.csv(rbind(cbind(vresult,yearclass="All"),vresult_sepyear),"vresult.csv")


# Compute confusion matrix
confmat <- table(pred_df_vote$predclass, pred_df_vote$refclass)
write.csv(confmat, "confmat.csv")
yearclass <- unique(pred_df_vote$yearclass)
y = yearclass[1]
pred_df_y <- pred_df_vote[pred_df_vote$yearclass==y,]
table(pred_df_y$predclass, pred_df_y$refclass)


# Create final model (use randomforest)
buildfinmodel <- function(traindata, predictors, rangermodel){
  mtry <- rangermodel$mtry
  set.seed(2021)
  td_b <- as.data.table(balance(traindata, "max", cat_col="Class"))
  set.seed(2022)
  grid_rf <- expand.grid(mtry = mtry)
  fit_rf <- trainControl(method = "none", seeds = 2022, classProbs = TRUE)
  set.seed(2023)
  m <- caret::train(td_b[,..predictors], td_b[["Class"]], method="rf", type="prob",
                    trControl = fit_rf, tuneGrid = grid_rf, 
                    importance = TRUE, verbose = TRUE)
  return(m)
}

rangermodel <- readRDS("D:/landsat/train/cv_pred.rds")
rf_model <- buildfinmodel(traindata, predictors, rangermodel)
saveRDS(rf_model, "D:/landsat/train/rf_model.rds")
